+++ lib/imap/backup/cli.rb
@@ -10,7 +10,6 @@ module Imap::Backup
     autoload :Backup, "imap/backup/cli/backup"
     autoload :Folders, "imap/backup/cli/folders"
     autoload :Local, "imap/backup/cli/local"
-    autoload :Mirror, "imap/backup/cli/mirror"
     autoload :Remote, "imap/backup/cli/remote"
     autoload :Restore, "imap/backup/cli/restore"
     autoload :Setup, "imap/backup/cli/setup"
@@ -21,6 +20,27 @@ module Imap::Backup
     include Helpers
 
     VERSION_ARGUMENTS = %w(-v --version).freeze
+
+    NAMESPACE_CONFIGURATION_DESCRIPTION = <<~DESC.freeze
+      Some IMAP servers use namespaces (i.e. prefixes like "INBOX"),
+      while others, while others concatenate the names of subfolders
+      with a charater ("delimiter") other than "/".
+
+      In these cases there are two choices.
+
+      You can use the `--automatic-namespaces` option.
+      This wil query the source and detination servers for their
+      namespace configuration and will adapt paths accordingly.
+      This option requires that both the source and destination
+      servers are available and work with the provided parameters
+      and authentication.
+
+      If automatic configuration does not work as desired, there are the
+      `--source-prefix=`, `--source-delimiter=`,
+      `--destination-prefix=` and `--destination-delimiter=` parameters.
+      To check what values you should use, check the output of the
+      `imap-backup remote namespaces EMAIL` command.
+    DESC
 
     default_task :backup
 
@@ -85,24 +105,7 @@ module Imap::Backup
 
       Some configuration may be necessary, as follows:
 
-      Some IMAP servers use namespaces (i.e. prefixes like "INBOX"),
-      while others, while others concatenate the names of subfolders
-      with a charater ("delimiter") other than "/".
-
-      In these cases there are two choices.
-
-      You can use the `--automatic-namespaces` option.
-      This wil query the source and detination servers for their
-      namespace configuration and will adapt paths accordingly.
-      This option requires that both the source and destination
-      servers are available and work with the provided parameters
-      and authentication.
-
-      If automatic configuration does not work as desired, there are the
-      `--source-prefix=`, `--source-delimiter=`,
-      `--destination-prefix=` and `--destination-delimiter=` parameters.
-      To check what values you should use, check the output of the
-      `imap-backup remote namespaces EMAIL` command.
+      #{NAMESPACE_CONFIGURATION_DESCRIPTION}
 
       Finally, if you want to delete existing emails in destination folders,
       use the `--reset` option. In this case, all existing emails are
@@ -160,7 +163,7 @@ module Imap::Backup
       If a folder list is configured for the SOURCE_EMAIL account,
       only the folders indicated by the setting are copied.
 
-      First, runs the download of the SOURCE_EMAIL account.
+      First, it runs the download of the SOURCE_EMAIL account.
       If the SOURCE_EMAIL account is **not** configured to be in 'mirror' mode,
       a warning is printed.
 
@@ -172,6 +175,11 @@ module Imap::Backup
     config_option
     quiet_option
     verbose_option
+    method_option(
+      "automatic-namespaces",
+      type: :boolean,
+      desc: "automatically choose delimiters and prefixes"
+    )
     method_option(
       "destination-delimiter",
       type: :string,
@@ -196,7 +204,7 @@ module Imap::Backup
     )
     def mirror(source_email, destination_email)
       non_logging_options = Imap::Backup::Logger.setup_logging(options)
-      Mirror.new(source_email, destination_email, **non_logging_options).run
+      Transfer.new(:mirror, source_email, destination_email, non_logging_options).run
     end
 
     desc "remote SUBCOMMAND [OPTIONS]", "View info about online accounts"